+++ fake-cucumber/javascript/src/EmptyPickleTestStep.ts
@@ -5,6 +5,8 @@ export default class EmptyPickleTestStep extends TestStep {
   public toMessage(): messages.TestStep {
     return {
       id: this.id,
+      stepDefinitionIds: [],
+      stepMatchArgumentsLists: []
     }
   }
 }